refactor(test): extract shared DEFAULT_OPTS into @pnpm/testing.command-defaults
12 command test suites had near-identical ~50-field DEFAULT_OPTS objects
copy-pasted between them. Extract the common fields into a single shared
package so each suite only declares its overrides.
