+++ deps/compliance/commands/test/licenses/utils/index.ts
@@ -1,53 +1,7 @@
-const REGISTRY = 'https://registry.npmjs.org'
+import { DEFAULT_OPTS as BASE_OPTS } from '@pnpm/testing.command-defaults'
 
 export const DEFAULT_OPTS = {
-  argv: {
-    original: [],
-  },
-  bail: true,
-  bin: 'node_modules/.bin',
-  ca: undefined,
-  cacheDir: '../cache',
-  cert: undefined,
-  excludeLinksFromLockfile: false,
-  extraEnv: {},
-  cliOptions: {},
-  fetchRetries: 2,
-  fetchRetryFactor: 90,
-  fetchRetryMaxtimeout: 90,
-  fetchRetryMintimeout: 10,
-  filter: [] as string[],
-  httpsProxy: undefined,
-  include: {
-    dependencies: true,
-    devDependencies: true,
-    optionalDependencies: true,
-  },
-  key: undefined,
-  linkWorkspacePackages: true,
-  localAddress: undefined,
-  lock: false,
-  lockStaleDuration: 90,
-  minimumReleaseAge: 0,
-  networkConcurrency: 16,
-  offline: false,
-  pending: false,
-  pnpmfile: ['./.pnpmfile.cjs'],
-  pnpmHomeDir: '',
-  preferWorkspacePackages: true,
-  proxy: undefined,
-  configByUri: {},
-  registries: { default: REGISTRY },
-  rootProjectManifestDir: '',
-  // registry: REGISTRY,
-  sort: true,
-  storeDir: '../store',
-  strictSsl: false,
-  userAgent: 'pnpm',
-  userConfig: {},
-  useRunningStoreServer: false,
-  useStoreServer: false,
+  ...BASE_OPTS,
+  registries: { default: 'https://registry.npmjs.org' },
   virtualStoreDir: 'node_modules/.pnpm',
-  workspaceConcurrency: 4,
-  virtualStoreDirMaxLength: process.platform === 'win32' ? 60 : 120,
 }
