+++ deps/compliance/commands/tsconfig.json
@@ -68,6 +68,9 @@
       "path": "../../../store/path"
     },
     {
+      "path": "../../../testing/command-defaults"
+    },
+    {
       "path": "../../../testing/mock-agent"
     },
     {
