+++ exec/commands/package.json
@@ -82,6 +82,7 @@
     "@pnpm/prepare": "workspace:*",
     "@pnpm/registry-mock": "catalog:",
     "@pnpm/test-ipc-server": "workspace:*",
+    "@pnpm/testing.command-defaults": "workspace:*",
     "@pnpm/workspace.projects-filter": "workspace:*",
     "@types/is-windows": "catalog:",
     "@types/ramda": "catalog:",
