+++ installing/commands/package.json
@@ -106,6 +106,7 @@
     "@pnpm/store.index": "workspace:*",
     "@pnpm/test-fixtures": "workspace:*",
     "@pnpm/test-ipc-server": "workspace:*",
+    "@pnpm/testing.command-defaults": "workspace:*",
     "@pnpm/testing.mock-agent": "workspace:*",
     "@pnpm/worker": "workspace:*",
     "@pnpm/workspace.projects-filter": "workspace:*",
