+++ installing/commands/tsconfig.json
@@ -106,6 +106,9 @@
       "path": "../../store/index"
     },
     {
+      "path": "../../testing/command-defaults"
+    },
+    {
       "path": "../../testing/mock-agent"
     },
     {
