+++ patching/commands/tsconfig.json
@@ -70,6 +70,9 @@
       "path": "../../store/path"
     },
     {
+      "path": "../../testing/command-defaults"
+    },
+    {
       "path": "../../workspace/project-manifest-reader"
     },
     {
