+++ releasing/commands/tsconfig.json
@@ -103,6 +103,9 @@
       "path": "../../resolving/resolver-base"
     },
     {
+      "path": "../../testing/command-defaults"
+    },
+    {
       "path": "../../workspace/projects-filter"
     },
     {
