+++ testing/command-defaults/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "@pnpm/testing.command-defaults",
+  "version": "0.0.0",
+  "private": true,
+  "description": "Shared DEFAULT_OPTS for pnpm command handler tests",
+  "keywords": [
+    "pnpm",
+    "pnpm11",
+    "testing"
+  ],
+  "license": "MIT",
+  "funding": "https://opencollective.com/pnpm",
+  "repository": "https://github.com/pnpm/pnpm/tree/main/testing/command-defaults",
+  "homepage": "https://github.com/pnpm/pnpm/tree/main/testing/command-defaults#readme",
+  "bugs": {
+    "url": "https://github.com/pnpm/pnpm/issues"
+  },
+  "type": "module",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "exports": {
+    ".": "./lib/index.js"
+  },
+  "files": [
+    "lib",
+    "!*.map"
+  ],
+  "scripts": {
+    "prepublishOnly": "pn compile",
+    "lint": "eslint \"src/**/*.ts\"",
+    "test": "pn compile",
+    "compile": "tsgo --build && pn lint --fix"
+  },
+  "dependencies": {
+    "@pnpm/registry-mock": "catalog:"
+  },
+  "engines": {
+    "node": ">=22.13"
+  },
+  "devDependencies": {
+    "@pnpm/testing.command-defaults": "workspace:*"
+  }
+}
